move funcs related to cached files to a single mod
